+++ Example/ManaKit/AppDelegate.swift
@@ -40,12 +40,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 maintainer.endActivity()
                 UserDefaults.standard.set(ManaKit.Constants.ScryfallDate, forKey: ManaKit.UserDefaultsKeys.ScryfallDate)
                 UserDefaults.standard.synchronize()
-                
+
                 ManaKit.sharedInstance.setupResources()
                 ManaKit.sharedInstance.configureTcgPlayer(partnerKey: "ManaGuide",
                                                           publicKey: "A49D81FB-5A76-4634-9152-E1FB5A657720",
                                                           privateKey: "C018EF82-2A4D-4F7A-A785-04ADEBF2A8E5")
-                
+
                 NotificationCenter.default.post(name: Notification.Name(rawValue: MaintainerKeys.MaintainanceDone),
                                                 object: nil,
                                                 userInfo: nil)